feat: :sparkles: Add Policies to User and Proc

+++ api/app/Http/Controllers/ProcsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Proc;
+use App\Models\User;
 use App\Http\Requests\ProcRequest;
 use Illuminate\Http\Request;
 
@@ -10,6 +11,7 @@ class ProcsController extends Controller
 {
     public function index()
     {
+        $this->authorize('viewAny', Proc::class);
         return Proc::with(['user', 'events'])
                    ->where('active', true)
                    ->get();
@@ -17,6 +19,7 @@ class ProcsController extends Controller
 
     public function store(ProcRequest $request)
     {
+        $this->authorize('create', Proc::class);
         $data = $request->validated();
         
         $data['number'] = $this->generateProcNumber();
@@ -44,11 +47,13 @@ class ProcsController extends Controller
 
     public function show(Proc $proc)
     {
+        $this->authorize('view', $proc);
         return $proc->load(['user', 'events.docs']);
     }
 
     public function update(ProcRequest $request, Proc $proc)
     {
+        $this->authorize('update', $proc);   
         $data = $request->validated();
         $proc->update($data);
         return response()->json($proc->load('user'));
@@ -56,6 +61,7 @@ class ProcsController extends Controller
 
     public function destroy(Proc $proc)
     {
+        $this->authorize('delete', $proc);
         $proc->delete();
         return response()->json(null, 204);
     }
